Fix bogus session ID specified for "local" op
The session ID specified was a randomly generated ID rather than the
actual session ID. This casued a 5 second delay for every "local" op
because of this code that tries to wait for the specified session:
https://github.com/moby/buildkit/blob/abde08a5531d809a395cf648a31bca932b009af0/source/local/local.go#L95-L101
